nitpicks(create-article): fix merge conflicts
