+++ src/AppRoutes.js
@@ -11,30 +11,22 @@ import Signin from './containers/SignIn/SignIn';
 import AllCategories from './containers/Categories/Categories';
 import ArticleCategory from './containers/Categories/SingleCategory';
 import NavBar from './components/NavBar';
-import Footer from './components/Footer';
 import CreateArticlePage from './containers/CreateArticle/CreateArticlePage';
 
 const Login = () => (
   <div><h1>Login page</h1></div>
 );
-import ProfilePage from "./containers/ProfilePage";
-import ArticlePage from "./containers/ArticlePage";
-import Homepage from "./containers/Home/Home";
-import Header from "./components/Header/Header";
-import Footer from "./components/Footer/Footer";
-import ProtectedRoute from "./containers/ProtectedRoute";
+// import ProfilePage from "./containers/ProfilePage";
+// import ArticlePage from "./containers/ArticlePage";
+// import Homepage from "./containers/Home/Home";
+// import Header from "./components/Header/Header";
+// import Footer from "./components/Footer/Footer";
+// import ProtectedRoute from "./containers/ProtectedRoute";
 // import NavBar from './components/NavBar';
-import CreateArticlePage from "./containers/CreateArticle/CreateArticlePage";
 
 const Sample = () => (
   <div className="landing-page">
     <h1>This is just a sample page for routes</h1>
-  </div>
-);
-
-const Login = () => (
-  <div>
-    <h1>Login Page</h1>
   </div>
 );
