+++ src/actionTypes/article.js
@@ -1,3 +1,3 @@
-export const FETCH_ARTICLE_SUCCESS = "FETCH_ARTICLE_SUCCESS";
-export const FETCH_ARTICLE_LOADING = "FETCH_ARTICLE_LOADING";
-export const FETCH_ARTICLE_ERROR = "FETCH_ARTICLE_ERROR";
+export const CREATE_ARTICLE_SUCCESS = "CREATE_ARTICLE_SUCCESS";
+export const CREATE_ARTICLE_FAILURE = "CREATE_ARTICLE_FAILURE";
+export const CREATE_ARTICLE_LOADING = "CREATE_ARTICLE_LOADING";
